Revert "Store a z-index id per-layer (#4128)"
This reverts commit 28a23372185c7a30c7790085a6a010c3cad6ad09, reversing
changes made to e992f84735a2ca166a7388d4eca69a6b4e6ee212.

+++ crates/gpui/src/styled.rs
@@ -12,7 +12,7 @@ pub trait Styled: Sized {
 
     gpui_macros::style_helpers!();
 
-    fn z_index(mut self, z_index: u16) -> Self {
+    fn z_index(mut self, z_index: u8) -> Self {
         self.style().z_index = Some(z_index);
         self
     }
